9624: Move all React components to top level
