+++ spec/javascript/components/search/SubmitterFilter.test.js
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 
 import { getFiltersStore } from './utils';
 
-import Component, { SUBMITTER_TYPES } from '../../../../app/javascript/components/search/SubmitterFilter';
+import Component, { SUBMITTER_TYPES } from '../../../../app/javascript/components/SubmitterFilter';
 
 const defaultProps = {
   filtersStore: getFiltersStore(),
