+++ spec/javascript/components/search/utils.js
@@ -1,6 +1,6 @@
 import FiltersModel from '../../../../app/javascript/components/search/FiltersModel';
 import ConditionSetModel from '../../../../app/javascript/components/ConditionSetModel/ConditionSetModel';
-import { submitterType } from '../../../../app/javascript/components/search/SubmitterFilter';
+import { submitterType } from '../../../../app/javascript/components/SubmitterFilter';
 
 export const getFiltersStore = () => new FiltersModel({
   conditionSetStore: new ConditionSetModel({
